Add a settings background image.
Add a saved() method to ui.validationForm and ui.input to notify the widgets
when a form has been successfully saved.  This allows the widgets to reset any
validation state used to determine whether or not the form can be submitted
(e.g. whether or not the input field has changed).