Updated to remove issues with the latest php5, primarily 'undefined index'.
To remove any remaining 'undefined index' warnings (many remain in the view
scripts and helpers), add '& ~E_NOTICE' to the 'error_reporting' setting in
your php.ini file.

Add 'name' to 'userAuth'

Fix background-image issues in Chrome for the star ratings.

Add an option to ui.validationForm.js to allow a caller to specify that the
submit button should NOT be disabled if the form is valid but there are no
changes from the initial values.  This is primarily to support auto-fill forms
(specifically for sign-on).

Fix remaining issues with user network.  This is a special "memberGroup" with
the properties:
    name:               'Network',
    ownerId:            %Id of user%,
    groupType:          'user',
    'controlMembers':   'owner',
    'controlItems':     'owner',
    'canTransfer':      false,
    'visibility':       'public' | 'private'